cellstudio-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cellstudio"
version = "0.2.0"
description = "A comprehensive deep learning framework for pathological cell analysis."
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "AlchemyAITech", email = "contact@alchemyai.com" }
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy<2.0.0",
    "opencv-python-headless>=4.7.0",
    "Pillow>=9.4.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
med = [
    "cellpose>=3.0.0",
    "segmentation-models-pytorch>=0.3.3",
]
yolo = [
    "ultralytics>=8.0.0",
]
timm = [
    "timm>=0.9.0",
    "transformers>=4.30.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
]

dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]
all = [
    "cellstudio[med,yolo,timm,viz,dev]",
]

[project.scripts]
cellstudio = "cellstudio.cli:main"

[tool.setuptools.packages.find]
include = ["cellstudio*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["cellstudio"]
